Rename 🚚: LayoutContent.jsx 파일 이름 및 폴더 변경 scroll, position:relative 추가
