+++ src/components/Layout/Layout.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { styled } from 'styled-components';
-import { COLOR } from '../utils';
+import { COLOR } from '../../utils';
 
 export default function Layout({ children }) {
     return <LayoutStyle>{children}</LayoutStyle>;
