+++ src/components/Layout/LayoutContent.jsx
@@ -1,13 +1,15 @@
 import React from 'react';
 import { styled } from 'styled-components';
-import { COLOR } from '../utils';
+import { COLOR } from '../../utils';
 
-export default function Content({ children }) {
+export default function LayoutContent({ children }) {
     return <ContentStyle> {children}</ContentStyle>;
 }
 
-const ContentStyle = styled.div`
+const ContentStyle = styled.main`
     flex-grow: 1;
     padding: 0 16px;
     background-color: ${COLOR.bgPrimaryColor};
+    position: relative;
+    overflow: scroll;
 `;
